fix delete product

See merge request anhemtn114/doantotnghiep!78

+++ app/Services/ProductService.php
@@ -3,6 +3,8 @@
 namespace App\Services;
 
 use App\Constants\Common;
+use App\Repositories\Contracts\ColorRepositoryInterface;
+use App\Repositories\Contracts\OrderItemsRepositoryInterface;
 use App\Repositories\Contracts\ProductReponsitoryInterface;
 use App\Services\Contracts\ProductServiceInterface;
 use Exception;
@@ -12,14 +14,24 @@ use Illuminate\Support\Facades\Log;
 
 class ProductService implements ProductServiceInterface
 {
-    protected ProductReponsitoryInterface $productReponsitory;
+    protected ProductReponsitoryInterface   $productReponsitory;
+    protected ColorRepositoryInterface      $colorRepositoryInterface;
+    protected OrderItemsRepositoryInterface $orderItemsRepository;
 
     /**
      * @param ProductReponsitoryInterface $repositoryInterface
+     * @param ColorRepositoryInterface $colorRepositoryInterface
+     * @param OrderItemsRepositoryInterface $orderItemsRepository
      */
-    public function __construct(ProductReponsitoryInterface $repositoryInterface)
+    public function __construct(
+        ProductReponsitoryInterface   $repositoryInterface,
+        ColorRepositoryInterface      $colorRepositoryInterface,
+        OrderItemsRepositoryInterface $orderItemsRepository,
+    )
     {
-        return $this->productReponsitory = $repositoryInterface;
+        $this->productReponsitory       = $repositoryInterface;
+        $this->colorRepositoryInterface = $colorRepositoryInterface;
+        $this->orderItemsRepository     = $orderItemsRepository;
     }
 
     /**
@@ -50,22 +62,63 @@ class ProductService implements ProductServiceInterface
      */
     public function create(array $attributes): mixed
     {
+        DB::beginTransaction();
         try {
-            if (isset($attributes['image_url'])) {
-                $image = $attributes['image_url'];
+           $attribute = $this->convertAttribute($attributes);
 
-                $attributes['image_url'] = handleImage($image);
-            } else {
-                $attributes['image_url'] = "no-image.png";
+            $product = $this->productReponsitory->create($attribute);
+
+            if (!$product) {
+                return null;
             }
 
-            $attributes['specifications'] = convertJson($attributes['specifications']);
+            $this->colorRepositoryInterface->create([
+                'product_id'   => $product['id'] ?? 0,
+                'color'        => $attributes['color'] ?? '',
+                'amount_color' => $attributes['amount'] ?? 0,
+            ]);
 
-            return $this->productReponsitory->create($attributes);
+            DB::commit();
+            return $product;
         } catch (Exception $exception) {
+            DB::rollBack();
             Log::error($exception->getMessage());
             return null;
         }
+    }
+
+    /**
+     * @param array $attributes
+     * @return array
+     */
+    private function convertAttribute(array $attributes): array
+    {
+        if (isset($attributes['image_url'])) {
+            $image = $attributes['image_url'];
+
+            $attribute['image_url'] = handleImage($image);
+        } else {
+            if ($attributes['oldImage']) {
+                $attributes['image_url'] = $attributes['oldImage'];
+            } else {
+                $attribute['image_url'] = "no-image.png";
+            }
+        }
+
+        $attribute['specifications'] = convertJson($attributes['specifications']);
+        $attribute['category_id']    = $attributes['category_id'] ?? 0;
+        $attribute['brand_id']       = $attributes['brand_id'] ?? 0;
+        $attribute['name']           = $attributes['name'] ?? '';
+        $attribute['price']          = $attributes['price'] ?? 0;
+        $attribute['old_price']      = $attributes['old_price'] ?? null;
+        $attribute['description']    = $attributes['description'] ?? null;
+        $attribute['tags']           = $attributes['tags'] ?? null;
+        $attribute['is_best_sell']   = $attributes['is_best_sell'] ?? 0;
+        $attribute['is_new']         = $attributes['is_new'] ?? 0;
+        $attribute['sort_order']     = $attributes['sort_order'] ?? 0;
+        $attribute['active']         = $attributes['active'] ?? 0;
+
+        return $attribute ?? [];
     }
 
     /**
@@ -76,17 +129,9 @@ class ProductService implements ProductServiceInterface
     public function update(array $attributes, int $id): mixed
     {
         try {
-            if (isset($attributes['image_url'])) {
-                $image = $attributes['image_url'];
+            $attribute = $this->convertAttribute($attributes);
 
-                $attributes['image_url'] = handleImage($image);
-            } else {
-                $attributes['image_url'] = $attributes['oldImage'];
-            }
-
-            $attributes['specifications'] = convertJson($attributes['specifications']);
-
-            return $this->productReponsitory->update($attributes, $id);
+            return $this->productReponsitory->update($attribute, $id);
         } catch (Exception $exception) {
             Log::error($exception->getMessage());
             return null;
@@ -100,10 +145,16 @@ class ProductService implements ProductServiceInterface
     public function delete(int $id): mixed
     {
         try {
+            $count = $this->orderItemsRepository->findWhere(['product_id' => $id])->count();
+            if ($count > Common::COUNT_DELETE) {
+                return null;
+            }
+
             $product = $this->productReponsitory->find($id);
 
             if ($product) {
                 $product->delete();
+                $this->colorRepositoryInterface->where(['product_id' => $id])->delete();
             }
 
             return $product;
